add Auth to wifiRess + more checks for modifying rights in scan

+++ Code/Java/Serveur/src/main/java/com/univtln/univTlnLPS/ressources/scan/WifiDataResources.java
@@ -1,11 +1,18 @@
 package com.univtln.univTlnLPS.ressources.scan;
 
 import com.univtln.univTlnLPS.dao.scan.WifiDataDAO;
+import com.univtln.univTlnLPS.model.administration.Administrateur;
+import com.univtln.univTlnLPS.model.administration.Superviseur;
+import com.univtln.univTlnLPS.model.scan.ScanData;
 import com.univtln.univTlnLPS.model.scan.WifiData;
+import com.univtln.univTlnLPS.security.annotations.JWTAuth;
+import jakarta.annotation.security.RolesAllowed;
 import jakarta.persistence.EntityTransaction;
 import jakarta.ws.rs.NotFoundException;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.SecurityContext;
 
 @Produces({MediaType.TEXT_XML, MediaType.APPLICATION_JSON})
 @Path("LaGarde")
@@ -24,11 +31,11 @@ public class WifiDataResources {
         }
     }
 
-    // add delete update
-
     @PUT
     @Path("wifis")
     @Consumes(MediaType.APPLICATION_JSON)
+    @RolesAllowed({"SUPER", "ADMIN"})
+    @JWTAuth
     public WifiData addWifiData(WifiData wifidata) throws IllegalArgumentException {
         if (wifidata.getId() != 0) throw new IllegalArgumentException();
 
@@ -46,8 +53,26 @@ public class WifiDataResources {
     @POST
     @Path("wifis/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
-    public WifiData updateWifiData(@PathParam("id") long id, WifiData wifidata) throws NotFoundException, IllegalArgumentException {
-        if (wifidata.getId() != id) throw new IllegalArgumentException();
+    @RolesAllowed({"SUPER", "ADMIN"})
+    @JWTAuth
+    public WifiData updateWifiData(@Context SecurityContext securityContext, @PathParam("id") long id, WifiData wifidata) throws NotFoundException, IllegalArgumentException {
+
+        // TODO check super has the rights to modify this wifi
+        if (wifidata.getId() != id && id == 0) throw new IllegalArgumentException();
+
+
+        // We check that super has the rights to modify this wifi (admin or supers wifiData creator)
+        if(!(securityContext.getUserPrincipal() instanceof Administrateur)) {
+
+            ScanData scanData = wifidata.getScanData();
+            if(scanData == null) throw new NotFoundException();
+
+            Superviseur superviseur = scanData.getSuperviseur();
+            if(superviseur == null) throw new NotFoundException();
+
+            if (!(securityContext.getUserPrincipal() ==  superviseur))
+                throw new IllegalArgumentException();
+        }
 
         try (WifiDataDAO wifiDataDAO = WifiDataDAO.of()) {
             EntityTransaction transaction = wifiDataDAO.getTransaction();
@@ -79,6 +104,8 @@ public class WifiDataResources {
 
     @GET
     @Path("wifis/size")
+    @RolesAllowed({"ADMIN"})
+    @JWTAuth
     public int getWifiDataSize() {
         try (WifiDataDAO wifiDataDAO = WifiDataDAO.of()) {
 
@@ -89,13 +116,35 @@ public class WifiDataResources {
 
     @DELETE
     @Path("wifis/{id}")
-    public void removeWifiData(@PathParam("id") long id) throws NotFoundException {
+    @RolesAllowed({"SUPER", "ADMIN"})
+    @JWTAuth
+    public void removeWifiData(@Context SecurityContext securityContext, @PathParam("id") long id) throws NotFoundException, IllegalArgumentException {
+
         try (WifiDataDAO wifiDataDAO = WifiDataDAO.of()) {
+
+            if(id == 0) throw new IllegalArgumentException();
+
+            WifiData wifiData = wifiDataDAO.find(id);
+            if( wifiData == null) throw new NotFoundException();
+
+            // We check that super has the rights to modify this wifi
+            // On verifie que l'utilisateur soit admin ou le superviseur du wifi du scan
+            if(!(securityContext.getUserPrincipal() instanceof Administrateur)) {
+
+                ScanData scanData = wifiData.getScanData();
+                if(scanData == null) throw new NotFoundException();
+
+                Superviseur superviseur = scanData.getSuperviseur();
+                if(superviseur == null) throw new NotFoundException();
+
+                if (!(securityContext.getUserPrincipal() ==  superviseur))
+                    throw new IllegalArgumentException();
+            }
+
             EntityTransaction transaction = wifiDataDAO.getTransaction();
 
             transaction.begin();
-            WifiData wifiData = wifiDataDAO.find(id);
-            if( wifiData == null) throw new NotFoundException();
+
             wifiDataDAO.remove(wifiData);
 
             transaction.commit();
@@ -104,6 +153,8 @@ public class WifiDataResources {
 
     @DELETE
     @Path("wifis")
+    @RolesAllowed({"ADMIN"})
+    @JWTAuth
     public void deleteWifiData() {
         try (WifiDataDAO wifiDataDAO = WifiDataDAO.of()) {
             wifiDataDAO.deleteAll();
